v1.5.30 -- After many tests, I ended up using a Makefile so the bumphunter example output will be included in the pkg

+++ R/derfinderReport.R
@@ -208,10 +208,13 @@ derfinderReport <- function(prefix, outdir = 'basicExploration',
     
     ## Locate Rmd if one is not provided
     if (is.null(template)) {
+        templateNull <- TRUE
         template <- system.file(
             file.path('basicExploration', 'basicExploration.Rmd'),
             package = 'regionReport', mustWork = TRUE
         )
+    } else {
+        templateNull <- FALSE
     }
     
     ## Load knitcitations with a clean bibliography
@@ -321,7 +324,7 @@ derfinderReport <- function(prefix, outdir = 'basicExploration',
             res <- render(paste0(output, '.Rmd'), output_format,
                 clean = .advanced_argument('clean', TRUE, ...))
         }
-        file.remove(paste0(output, '.Rmd'))
+        if(templateNull) file.remove(paste0(output, '.Rmd'))
     
         ## Open
         if (browse) browseURL(res)
